Rename audit to event
