+++ management/server/event/sqlite.go
@@ -1,4 +1,4 @@
-package audit
+package event
 
 import (
 	"database/sql"
